+++ cookbook/models/meta/llama/basic_stream.py
@@ -1,10 +1,8 @@
 from typing import Iterator  # noqa
 from agno.agent import Agent, RunResponse  # noqa
-from agno.models.llama import LlamaOpenAI
+from agno.models.meta import Llama
 
-agent = Agent(
-    model=LlamaOpenAI(id="Llama-4-Maverick-17B-128E-Instruct-FP8"), markdown=True
-)
+agent = Agent(model=Llama(id="Llama-4-Maverick-17B-128E-Instruct-FP8"), markdown=True)
 
 # Get the response in a variable
 # run_response: Iterator[RunResponse] = agent.run("Share a 2 sentence horror story", stream=True)
