+++ cookbook/models/meta/llama/knowledge.py
@@ -2,7 +2,7 @@
 
 from agno.agent import Agent
 from agno.knowledge.pdf_url import PDFUrlKnowledgeBase
-from agno.models.llama import Llama
+from agno.models.meta import Llama
 from agno.vectordb.pgvector import PgVector
 
 db_url = "postgresql+psycopg://ai:ai@localhost:5532/ai"
