+++ cookbook/models/meta/llama/tool_use.py
@@ -1,13 +1,12 @@
 """Run `pip install duckduckgo-search` to install dependencies."""
 
 from agno.agent import Agent
-from agno.models.llama import Llama
+from agno.models.meta import Llama
 from agno.tools.duckduckgo import DuckDuckGoTools
 
 agent = Agent(
     model=Llama(id="Llama-4-Maverick-17B-128E-Instruct-FP8"),
     tools=[DuckDuckGoTools()],
-    show_tool_calls=True,
     markdown=True,
 )
-agent.print_response("Whats happening in France?", stream=True)
+agent.print_response("Tell me about Llama API.")
