+++ cookbook/models/meta/llama/tool_use_stream.py
@@ -1,7 +1,7 @@
 """Run `pip install duckduckgo-search` to install dependencies."""
 
 from agno.agent import Agent
-from agno.models.llama import Llama
+from agno.models.meta import Llama
 from agno.tools.duckduckgo import DuckDuckGoTools
 
 agent = Agent(
@@ -10,4 +10,4 @@ agent = Agent(
     show_tool_calls=True,
     markdown=True,
 )
-agent.print_response("Whats happening in France?")
+agent.print_response("Whats happening in France?", stream=True)
