+++ cookbook/models/meta/llama_openai/basic_stream.py
@@ -1,8 +1,10 @@
 from typing import Iterator  # noqa
 from agno.agent import Agent, RunResponse  # noqa
-from agno.models.llama import Llama
+from agno.models.meta import LlamaOpenAI
 
-agent = Agent(model=Llama(id="Llama-4-Maverick-17B-128E-Instruct-FP8"), markdown=True)
+agent = Agent(
+    model=LlamaOpenAI(id="Llama-4-Maverick-17B-128E-Instruct-FP8"), markdown=True
+)
 
 # Get the response in a variable
 # run_response: Iterator[RunResponse] = agent.run("Share a 2 sentence horror story", stream=True)
