+++ cookbook/models/meta/llama_openai/metrics.py
@@ -1,7 +1,7 @@
 from typing import Iterator
 
 from agno.agent import Agent, RunResponse
-from agno.models.llama import LlamaOpenAI
+from agno.models.meta import LlamaOpenAI
 from agno.tools.yfinance import YFinanceTools
 from agno.utils.pprint import pprint_run_response
 from rich.pretty import pprint
