+++ libs/agno/agno/models/meta/llama.py
@@ -38,7 +38,7 @@ class Llama(Model):
     id: str = "Llama-4-Maverick-17B-128E-Instruct-FP8"
     name: str = "Llama"
     provider: str = "Llama"
-    supports_native_structured_outputs: bool = True
+    supports_native_structured_outputs: bool = False
 
     # Request parameters
     max_tokens: Optional[int] = None
